simplify SELinux labeling
This commit moves the SELinux label code after the tree is sorted and
the inode table is generated. Sorting helps to make sure that the tree
will always be traversed in a defined, deterministic order and likewise
the creation of xattrs happens in a defined, deterministic order.

Second, we can now use the inode table instead of having to implement a
recursive tree traversal yet again.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>