Cleanup: split fstree sort into 2 fstree independend functions
Make tree node list sort and recursive variant available and independend
of the fstree_t.

This is considered cleaner, since the fstree_t actually isn't needed for
any of this and we can just call the recusvie sort on the root instead,
and we can use the sort implementation directly for things like the
upcoming unit test.

Also this commit splits up the merge/sort implementation into a seperate
split and merge functions to make the code somewhat more readable.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>