Fix inode fragment & sparse counter initialization
The sparse attribute should be initialized with 0. It is subtracted
from the file size in the kernel to determine the file size on disk
to report to stat(2). So it actually didn't matter anyway (just for
correctness sake).

Second, the fragment index and offset should be set to 0xFFFFFFFF in
the rare case that no fragments exist.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>