Fix double free in GNU tar sparse file parser
The intention was to free the temporary object in addition to the list.
Since the temp pointer is also used for iterating the list, this resulted
in a double free instead of the intended.

This commit fixes the double free by replacing the recycled variable based
list free with the helper function intended to do this.

Bug found using scan-build.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>