tar writer: replace PAX headers with GNU extensions
Some experiments seem to indicate that the various GNU extensions are
more widely supported than their POSIX equivalents[1]. Possibly because
they are easier to implement and possibly because of the wide spread
use of GNU tar.

This commit replaces the PAX writer in the write_tar_header implementation
with a GNU extension based writer.

The writer is also cleaned up by removing all global state. The record
counter is moved outside into the tar2sqfs program and passed in as
function argument.

[1] https://dev.gentoo.org/~mgorny/articles/portability-of-tar-features.html

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>