Simplify writer code for files without a fragment
The data writer sparse block code can take advantage that it can add a
block size instead of a fragment and doesn't have to initialize the
framgent location.

In return, the tree node to inode serialization code doesn't need a
special case for sparse file anymore and can now also handle files that
are forced to not have a fragment.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>