Add a way to optionally keep the original time stamps
First of all, this commit adds a mod_time field to a tree node. When
creating the tree node, the field is set from the struct stat. When
scanning a directory, the time stamps from the input are used if set.

Second, the libsqfs code that reads inodes is modified to store the
mod_time from the inode in the fstree node and to write the tree node
into a generated inode.

Finally, tar2sqfs is modified to optionally keep the timestamps from
the tar archive instead of setting defaults. gensquashfs is similarly
modified to keep the input timestamps if specified.

The result is as follows:
 - sqfs2tar will always carry the timestamps from the squashfs over
   to the tar ball.
 - tar2sqfs will set defaults, unless explicitly asked to preserve
   the mtime from the tar ball.
 - gensquashfs can optionally preserve the mtime from the input
   hierarchy it processes if only --pack-dir is specified.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>