Generate linear file list in fstree
Instead of doing DFS on the fly in gensquashfs, churn out a linked list
of all files in an archive.

Future improvements in packing strategies can go into this file.

This can also be usefull for other purposes in the future, such as file
deduplication or as a work queue for the unpacker.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>