cleanup: remove atime/ctime processing code
This commit removes all the code for parsing and processing atime/ctime
and values and related test code.

Caring about those is kind of pointless because squashfs can only store
mtime in inodes. The only relevant place is when generating a struct
stat from a squashfs inode or an fstree node.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>