Terminate the unpacker child processes if the main process exits
Should the main process exit (e.g. the user presses CTRL+C) or explcitily
sends it a signal, the desired behaviour is for the children to stop
unpacking and exit.

This commit adds a line to configure the kernel to send SIGKILL to the
children if their parent dies.

The same option also exists on various BSDs (not all of them) but with a
different name which has to be checked for and adjusted should the program
be required to run on one of those.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>