Add a helper utility to compare filesystem trees
The intended use case is to compare two mounted or unpacke squashfs
images, so a repacked test image can be compared against its original
or an image unpacked with unsquashfs can be compared with an image
unpacked by rdsquashfs or sqfs2tar.

Since the tool is only intended to aid development (specifically
automated testing), it is not installed by `make install`.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>