cleanup: unify all the code that reads squashfs images
This commit creates a new data structure called 'sqfs_reader_t' that
takes care of all the repetetive tasks like opening the file, reading
the super block, creating the compressor, deserializing an fstree and
creating a data reader.

This in turn makes it possible to remove all the duplicate code from
rdsquashfs and sqfs2tar.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>