Tune the paranoia down a bit
size_t is guaranteed to be large enough to measure the size of things in
memory, so when doing exactely that (e.g. strlen(a) + strlen(b)), checking
for overflow is pointless since both objects are already in memory. If the
addition would overflow, the two strings would occupy more memory than
addressable.

(Possible exception being some kind of harward style architecture with
the two strings being in different kinds of memory of course.)

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>