Split libsquashfs.a into low seperate libraries
The idea is to make libsquashfs.a independend of libfstree.a, so it becomes
a general purpose squashfs manipulation library. All the high level glue code
for libfstree.a and utilites that are overly specific with to tools are moved
to a seperate librarby.

This commit makes the first step by moving the stuff with dependencies on
libfstree to a seperate library.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>