Move command line processing stuff out of compressor code
This commit moves stuff like printing help text, command line option
processing and enumerating available processors on stdout out of
the generic compressor code.

The option string is replaced with a structure that directly exposese
the tweakable parameters for all compressors. A function for parsing
the command line arguments into this structure is added in sqfshelper.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>