Replace file descriptor IO in data_writer with sqfs_file_t
First, this commit moves the create-blocks-from-fd function over to
libsquashfs and ports it to work on an sqfs_file_t instead.

Second, the function in the data_writer that reads from a file descriptor
is adjusted to use an sqfs_file_t instead.

Finally, the tools that use it have to be adjusted accordingly.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>