Add directory reader data structure
This moves a lot of the stuff that is done manually in the tree
deserializer to a generic helper in libsquashfs.

Due to how the fstree is implemented, as a work around, the inode needs
to be temporarily stored in the tree node, but some of the directory
details could be removed.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>