Remove use of ssize_t from library API
ssize_t is only available on POSIX platforms and even there it is
only defined to hold at least -1 in the range of negative numbers.

This commit replaces ssize_t return types with sqfs_s32 and the
coresponding function arguments with sqfs_u32. Because the range
of positiv numbers for a signed 32 bit number is only half that of
the unsigned version, additional checks have to be added.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>