Store directory index in generic inode structure
This commit modifies the dir writer to store the directory index directly
in the payload area of the inode it creates. The size of the index in bytes
is stored in a seperate field. Storing the index is moved to the write
inode function.

The read inode function is in turn also modified to read the directory
index from disk and actually store it in the inode structure in memory
for later use.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>