Fix tar2sqfs: check for empty names or './' entrty
This commit makes sure entries with empty file names are
skipped by tar2sqfs.

The case of a directory with the name './' is treated specially,
because tar can generate in some common use cases.

For instance:

$ mkdir foo
$ touch foo/bar
$ tar -C ./foo -c . > out.tar
$ tar tf out.tar

will generate a tar file with an entry named './'.

For the future it might be interesting to turn that into the root
inode somehow.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>