Fix: Move LZO compressor from libsquashfs to libcommon
The liblzo2 library is licensed under GPLv2, so it is not possible to
distribute binaries of libsquashfs that link against liblzo2 under
LGPL.

This commit moves the LZO compressor implementation to libcommon,
where this isn't a problem, since the tools themselves are licensed
under GPLv3.

It removes the ability of libsquashfs to read or generate LZO compressed
SquashFS images, but the tools still can.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>