Only check for OS specific bad filenames when unpacking
When converting a SquashFS image to a tarball, it makes no sense to
refuse conversion if the filename is considered evil by the OS.

This patch adds an option to is_filename_sane to check if the OS has
a problem with the given file name. sqfs2tar sets it to false and
converts everything while rdsquashfs sets it to true when unpacking.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>