Fix out of bounds writes in lzo compressor
The liblzo compression functions don't do any bounds checking,
instead they expect the destination buffer to be large enough
to hold the worst case encoding.

This commit modifies the lzo compressor to use a scratch buffer
for compression (with worst case size) and only copy to the actual
destination if the result fits.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>