Add a simplistic integration and regression test
This test basically consists of the Canterbury Corpus which is turned
into a SquashFS image using every supported compressor, with every
supported block size, with and without tail end packing.

The results MUST have an exact, given sha512sum. If that changes, it
means either a regression in the output format, something broke or
non-deterministic packing behaviour.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>