Always append '/' to names of directories
In the distant past, the first versions of tar did not have any header
field to indicate the type of the file being packed. Instead, they used
to append a '/' to the filename to indicate that an entry is a actually
a directory. GNU tar still does this today, despite having a header
field for file type.

This commit modifies sqfs2tar to append slashes to directory names
for compatibillity and to mimic GNU tar behaviour more closely.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>