Cleanup: remove the payload pointers from sqfs_inode_generic_t
There are 3 types of extra payload:
 - Directory index
 - File block sizes
 - Symlink target

This commit removes the type specific pointers and modifies the code
to use the payload area directly. To simplify the file block case and
mitigate alignment issues, the type of the extra field is changed to
sqfs_u32.

For symlink target, the extra field can simply be cast to a character
pointer (it had to be cast anyway for most uses). For block sizes,
probably the most common usecase, it can be used as is. For directory
indices, there is a helper function anyway.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>