Split the block writing/deduplication away from the block processor
This commit moves the entire block writing and deduplication of data
blocks over to a different data type named "block writer".

For simplicity, the interfaces of the block processor are left as is
and are turned into warppers. Likewise, most of the code in the block
writer is just verbatim from the block processor, to be cleaned up in
subsequent commits.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>