Cleanup: move utilities back out of libsquashfs
This commit removes the allocation helpers and string table functions
out of libsquashfs back into a "libutil.a". The problem of libsquashfs
exporting stuff that it shouldn't is resolved by retaining the internal
attributes and directly adding the source to libsquashfs instead of
trying to somehow link against libutil.la.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>