Make hard link detection scale better
Instead of doing a linear search, which scales quadratically, use a
red-black tree with inode numbers as key for finding hard links.

To reiterate: we can't just use a flat table, because the SquashFS
file is potentially untrusted and the inode numbers can be anything,
no matter what the super block says.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>