Cleanup: match xattr reader API closer to id table API
Instead of creating everything in the "create" function, cleanup and
create/initialize stuff in a "load" function. This allows the xattr
reader to be reset/re-used and adds the benefit of not having to
lug around references to the super block, compressor and file (altough
the later two are hidden inside the meta reader).

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>