Restore workaround for unaligned reads in xxhash
The code was originally used inside the block processor, where 32 bit
aligned data could be guaranteed. If it is available in libutil, I
cannot possibly guarantee for alignment in future use elsewhere. Even
for the block processor it was rather risky "remember this detail very
well" buisness.

This commit restores the unaligned read treatment of the original.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>