Do not try to read back the compressor options
None of the currently implemented compressors do anything with that data.
They are all at the mercy of the data actually in the image.

This commit removes the code from sqfs2tar and rdsquashfs that decodes
the options, which also has the side effect of increasing compatibillity
with some non-confirming images.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>