Skip PAX global headers
Tar archives can contain set two kinds of PAX headers:
 - local headers that modify the attributes of the next file
 - global headers that set defaults for all files

The later is used "... not widely used", according to tar(5)
and has been deliberately not implemented.

Some programs (e.g. git-archive) *do* generate them (in the case
of git, it stores the commit hash).

This commit adds a code path that skips a PAX global header entirely
and resumes tar parsing, instead of erroneusly reporting it as an
entry.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>