Enable uint128_t path
I forgot to enable this when I copied it over from Mesa. Mesa's
meson configuration system checks that a C program using the uint128_t
type compiles, but I think this is likely unnecessary. Simply check the
macro that clang and gcc define.

This cuts the .text size of hash_table.o by 160 bytes or about 4% on my
system.

Signed-off-by: Matt Turner <mattst88@gmail.com>