Cleanup/fix: gensquashfs: split directory scanning from xattr scanning
On the one hand, this commit cleanes the code a bit by splitting
the "scan directory contents" code from the "scan xattrs from
directory contents" and moving the later in a seperate file.

On the other hand, the xattr scanning is now done *after* the fstree
is post processed, which includes sorting it. This way, the xattrs
are always added in a deterministic, reproducible order.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>