Add a simple test script for the gensquashfs packdir + allroot use case
Since this is a fairly common use case, it deserves a simple test case
to check out that e.g. option processing hasn't been botched up (again).

As input directory, the licenses directory is used as it contains no
intermediate build output and should change fairly infrequently.

The test is enabled irregardless of the corpora-test option.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>