block processor: don't zero initialize the block payload area
In the block processor, the payload area is only accessed up to
the indicated size. Even the part that is accessed is initialized
by copying data into the block before increasing the size, so there
is no real point in zero-initializing hundres of kilobytes if not
megabytes of payload area, especially since this is done in the
locked, serial path of the block processor.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>