sqfsdiff: extract compressor options, but don't fail on error
This commit modifies sqfsdiff to extract the compressor options from
the squashfs image and store them in a compressor configuration if
possible. The failure path is modified to *not* burst into flames on
error, because those options are not required by any compressor to
read data from the disk and pretty much every vendor modifed SquashFS
has messed with those to the point that they cannot be propperly
decoded (or the flag is set and there are no options).

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>