Cleanup: clear naming confusion lzma vs xz in some comments
Previously, the XZ compressor was called LZMA at some points in the
code, because it actually referres to LZMAv2 and the older LZMAv1
wasn't implemented. Since LZMAv1 is supported for some time now,
simply calling it LZMA is confusing.

This commit tries to replace the last bits where XZ is referred to
as LZMA, which turns out to be only a hand full of comments.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>