Fix block bounds checking in libsquashfs data reader
Instead of doing the fragile size comparison in both loops, simply
bail from the function if offset is out of bounds, clamp the size
to the available range of the file and abail if it is zero.

As a result, a lot of checks can be removed and the function will not
return data beyond EOF.

This problem occoured with files that have a short last block instead
of a fragment.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>