Cleanup: move zlib/lz4 code from lib/sqfs/comp/ to lib/
The source code of a modified liblz4 and zlib are included with the
option to compile them into libsquashfs if they are not available on
the system.

So far, the source code was included directly in the compressor sub
directory within libsqsuashfs. This commit moves the libraries out
into the lib directory.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>