Fix block processor single block with don't fragment flag bug
This commit fixes a bug where the block processor state machine would
not add the "last block" flag if there is only one not entirely filled
block and the "don't fragment" flag is set. If the flag isn't set, the
inode start block position is not updated and points to the beginning
of the image instead.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>