Fix nonexistant gnu tar sparse format 1.0 support
Contrary to previous claims, support for the GNU tar sparse format 1.0
was missing entirely (the newest of their 3 different sparse mapping
formats). This oversight wasn't caught, because the unit test was
compiling the wrong source file and tar2sqfs had no problem processing
the test file because it is still a valid POSIX-ish tar archive (but
the sparse part was missing and the mapping embedded in the file).

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>