Cleanup: reduce tar test cases to a few generic C files
This commit removes the existing tar test cases that simply call the
generic test case function with several different paths with generic
test case source files that are parameneterized via the pro-processor.
For each tar archive, a separate test case is generated.

On the one hand, this reduces the test source code to practically
nothing. On the other hand, a test binary is generated for every
distinct test case, instead of one per group and we get more detailed
insights if something goes wrong.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>