Add a check whether libzstd supports stream compression
Streaming compression was added fairly recently to zstd (and then
the API was changed a few times).

Rather than rely on libzstd versioning macros, this commit adds an
m4 script to test at configure time if the enums/functions we need
are available by trying to compile a small sample.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>