Fix: rdsquashfs: describe: don't escape the prefixed file input path
If rdsquashfs describes a filesystem, is configured to use a prefix-path
and a file contains a space, the resulting input path has the prefix
printed as is and the rest of the string in quotation marks.

gensquashfs simply takes the entire rest of the line as is as its
input path and no longer finds the file.

Fix this by omitting the escapes and quotation marks for the
input path.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>