Remove package scripts from release tarballs
Some of the build scripts contain hashes of the release tarballs. When
releasing a new version, in addition to bumping all the version numbers,
this would require the amazing stunt of putting the hashes of the
release tarball into the release tarball itself, for a number of
different hash algorithms.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>